requirements.txt
Django==2.1.15
django-geoposition-2==0.3.11
Pillow==7.2.0
pytz==2020.1
